backend/requirements-full.txt
# Full dependency set — used once the ML pipeline lands on Day 8.
# Kept here so we don't lose the resolved versions when slimming down
# `requirements.txt` for the OAuth-only stage of the project.
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
Authlib==1.7.0
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
cloudpickle==3.1.2
cryptography==47.0.0
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
joblib==1.5.3
joserfc==1.6.4
llvmlite==0.47.0
numba==0.65.1
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pluggy==1.6.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pywebpush==2.0.3
Pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
ruff==0.15.12
scikit-learn==1.8.0
scipy==1.17.1
shap==0.51.0
six==1.17.0
slicer==0.0.8
starlette==1.0.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
